Large Scale Refactor
In order to get this project up to speed and into a good base point for future work, I've reorganized some of the architecture and underlying behaviour.

The largest change is that now constant buffers are owned by shaders and bound in a large 13-wide batch. In future revisions, this may change substantially.

I've also refactored much of the existing naming style to remove the underscore.